requirements.txt
streamlit==1.38.0
web3==7.2.0
Pillow==10.4.0
imagehash==4.3.1
requests==2.32.3
python-dotenv==1.0.1
numpy
scipy
gitpython